[chore] 폰트 익스텐션 수정
