+++ Melon_iOS/Melon_iOS/Global/Extensions/UILabel+.swift
@@ -1,0 +1,34 @@
+//
+//  UILabel+.swift
+//  Melon_iOS
+//
+//  Created by 조영서 on 11/16/25.
+//
+
+import UIKit
+
+extension UILabel {
+
+    func setPretendard(_ style: UIFont.PretendardStyle) {
+        self.font = .pretendard(style)
+        applyTypography(style)
+    }
+
+    private func applyTypography(_ style: UIFont.PretendardStyle) {
+        guard let text = self.text else { return }
+
+        let attributed = NSMutableAttributedString(string: text)
+
+        let paragraph = NSMutableParagraphStyle()
+        paragraph.minimumLineHeight = style.lineHeight
+        paragraph.maximumLineHeight = style.lineHeight
+
+        attributed.addAttributes([
+            .font: UIFont.pretendard(style),
+            .kern: style.letterSpacing,
+            .paragraphStyle: paragraph
+        ], range: NSRange(location: 0, length: attributed.length))
+
+        self.attributedText = attributed
+    }
+}
